Agent Web/.history/requirements_20250215172700.txt
streamlit>=1.32.0
langchain>=0.1.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
duckduckgo-search>=4.4.3
langchain-google-genai>=0.0.6 